Move UnifiedConfig into the config package.
