+++ confgenerator/testdata/invalid/windows/all-not_in_yaml_format/golden_error
@@ -1,2 +1,2 @@
 the agent config file is not valid YAML. detailed error: yaml: unmarshal errors:
-  line 1: cannot unmarshal !!str `sdfd` into confgenerator.UnifiedConfig
+  line 1: cannot unmarshal !!str `sdfd` into config.UnifiedConfig
